+++ photutils/psf/__init__.py
@@ -4,7 +4,8 @@ This subpackage contains modules and packages for point spread function
 photometry.
 """
 
-from .building import *      # noqa
+from .epsf import *          # noqa
+from .epsf_stars import *    # noqa
 from .funcs import *         # noqa
 from .groupstars import *    # noqa
 from .matching import *      # noqa
