--- photutils/psf/building/__init__.py
@@ -1,8 +1,0 @@
-# Licensed under a 3-clause BSD style license - see LICENSE.rst
-"""
-This subpackage contains modules and packages to build point spread
-functions.
-"""
-
-from .epsf import *
-from .stars import *
